Update README to include links to community libraries
From https://github.com/athenahealth/mdp/blob/master/community/README.md: If you have a snippet or module you'd like to share, submit a pull request for one of these language folders and we'll add it! If your sample code is substantial enough to merit a separate repo, consider creating your own and letting us know so we can link to it on the main README page.